+++ tools/apiview/emitters/cadl-apiview/src/apiview.ts
@@ -34,7 +34,6 @@ import { ApiViewDiagnostic, ApiViewDiagnosticLevel } from "./diagnostic.js";
 import { ApiViewNavigation } from "./navigation.js";
 import { generateId, NamespaceModel } from "./namespace-model.js";
 import { LIB_VERSION } from "./version.js";
-import { type } from "os";
 
 const WHITESPACE = " ";