+++ tools/apiview/emitters/cadl-apiview/src/emitter.ts
@@ -5,11 +5,14 @@ import {
   getServiceNamespace,
   getServiceTitle,
   getServiceVersion,
+  Namespace,
   NoTarget,
   Program,
+  ProjectionApplication,
+  projectProgram,
   resolvePath,
 } from "@cadl-lang/compiler";
-import { buildVersionProjections } from "@cadl-lang/versioning";
+import { buildVersionProjections, getVersion, getVersionsForEnum, VersionMap } from "@cadl-lang/versioning";
 import path from "path";
 import mkdirp from "mkdirp";
 import { ApiView } from "./apiview.js";
@@ -20,8 +23,9 @@ const defaultOptions = {
 } as const;
 
 export interface ResolvedApiViewEmitterOptions {
-  outputFile: string;
-  serviceNamespace?: string;
+  outputPath: string;
+  namespace?: string;
+  version?: string;
 }
 
 export async function $onEmit(program: Program, emitterOptions?: ApiViewEmitterOptions) {
@@ -37,12 +41,73 @@ export function resolveOptions(
   const resolvedOptions = { ...defaultOptions, ...options };
 
   return {
-    outputFile: resolvePath(
+    outputPath: resolvePath(
       resolvedOptions["output-dir"] ?? `${program.compilerOptions.outputDir!}/apiview`,
       resolvedOptions["output-file"]
     ),
-    serviceNamespace: resolvedOptions["service-namespace"]
+    namespace: resolvedOptions["namespace"],
+    version: resolvedOptions["version"],
   };
+}
+
+function resolveServiceVersion(program: Program): string | undefined {
+  // FIXME: Fix this wonky workaround when getServiceVersion is fixed.
+  const value = getServiceVersion(program);
+  return value == "0000-00-00" ? undefined : value;
+}
+
+function resolveNamespaceString(program: Program, namespace: Namespace): string | undefined {
+  // FIXME: Fix this wonky workaround when getNamespaceString is fixed.
+  const value = program.checker.getNamespaceString(namespace);
+  return value == "" ? undefined : value;
+}
+
+// TODO: Up-level this logic?
+function resolveAllowedVersions(program: Program, namespace: Namespace): string[] {  
+  const allowed: string[] = [];
+  const serviceVersion = resolveServiceVersion(program);
+  const versions = getVersion(program, namespace)?.getVersions();
+  if (serviceVersion != undefined && versions != undefined) {
+    throw new Error("Cannot have serviceVersion with multi-API.");
+  }
+  if (serviceVersion != undefined) {
+    allowed.push(serviceVersion);
+  } else if (versions != undefined) {
+    for (const item of versions) {
+      allowed.push(item.name);
+    }
+  } else {
+    throw new Error("Unable to resolve allowed versions.");
+  }
+  return allowed;
+}
+
+function resolveVersionValue(program: Program, namespace: Namespace, version: string): string {
+  const versions = getVersion(program, namespace)!.getVersions();
+  return versions.filter((item) => item.name == version).map((item) => item.value)[0];
+}
+
+function resolveProgramForVersion(program: Program, namespace: Namespace, versionKey: string): Program {
+  const version = resolveVersionValue(program, namespace, versionKey);
+  const projections = buildVersionProjections(program, namespace).filter((item) => item.version == version);
+  if (projections.length == 0) {
+    // non-multi-version scenario. Return original program.
+    return program;
+  } else {
+    // alias could result in an api version appearing twice, so always take the first
+    const projection = projections[0];
+    projection.projections.push({
+      projectionName: "atVersion",
+      arguments: [version],
+    });
+    const commonProjections: ProjectionApplication[] = [
+      {
+        projectionName: "target",
+        arguments: ["json"],
+      },
+    ];
+    return projectProgram(program, [...commonProjections, ...projection.projections]);
+  }
 }
 
 function createApiViewEmitter(program: Program, options: ResolvedApiViewEmitterOptions) {
@@ -53,42 +118,46 @@ function createApiViewEmitter(program: Program, options: ResolvedApiViewEmitterO
     if (!serviceNs) {
       throw new Error("No namespace found");
     }
-    // TODO: Supply `version` option to select specific version
-    let versionString = getServiceVersion(program);
-    // TODO: Fix this wonky workaround when getServiceVersion is fixed.
-    if (versionString == "0000-00-00") {
-      versionString = undefined;
-    }
-    const namespaceString = options.serviceNamespace ?? program.checker.getNamespaceString(serviceNs);
-    if (namespaceString == "") {
+    const versionString: string = options.version ?? resolveServiceVersion(program) ?? "latest";
+    const namespaceString = options.namespace ?? resolveNamespaceString(program, serviceNs);
+    if (namespaceString == undefined) {
       reportDiagnostic(program, {
         code: "use-namespace-option",
         target: NoTarget
       });
+      return;
     }
-    const versions = buildVersionProjections(program, serviceNs);
-    if (versions.length) {
-      // TODO: Add heuristic to choose "latest" if not otherwise supplied
-      // TODO: Validate that if version is supplied, it is found in the versioning list. Otherwise error.
+    const allowedVersions = resolveAllowedVersions(program, serviceNs);
+    if (allowedVersions.filter((version) => version == versionString).length == 0) {
+      reportDiagnostic(program, {
+        code: "version-not-found",
+        target: NoTarget,
+        format: {
+          version: versionString,
+          allowed: allowedVersions.join(" | "),
+        },
+      })
+      return;
     }
-    // FIXME: Remove this weird workaround when this call is fixed.
+    // FIXME: Fix this wonky workaround when getServiceTitle is fixed.
     let serviceTitle = getServiceTitle(program);
     if (serviceTitle == "(title)") {
       serviceTitle = namespaceString;
     }
+    const resolvedProgram = resolveProgramForVersion(program, serviceNs, versionString);
     const apiview = new ApiView(serviceTitle, namespaceString, versionString);
-    apiview.emit(program);
+    apiview.emit(resolvedProgram);
     apiview.resolveMissingTypeReferences();
 
     if (!program.compilerOptions.noEmit && !program.hasError()) {
-      const outputFolder = path.dirname(options.outputFile);
+      const outputFolder = path.dirname(options.outputPath);
       try {
         await mkdirp(outputFolder);
       } catch {
         // mkdirp fails during tests
       }
       await emitFile(program, {
-        path: options.outputFile,
+        path: options.outputPath,
         content: JSON.stringify(apiview.asApiViewDocument()) + "\n"
       });  
     }