+++ tools/apiview/emitters/cadl-apiview/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 0.1.2 (Unreleased)
 No longer suppress `@doc`, `@summary`, and `@example` decorators. These can be toggled using the APIView UI.
 Support rendering multi-line strings.
+Change default path for generating artifacts.
 
 ## Version 0.1.1 (10-13-2022)
 Support compiler-level noEmit option.