+++ tools/apiview/emitters/cadl-apiview/CHANGELOG.md
@@ -1,13 +1,15 @@
 # Release History
 
 ## Version 0.1.2 (Unreleased)
+Support `service-namespace` emitter option to filter the appropriate namespace when it cannot be automatically resolved.
+  This is primarily intended for creating APIViews for libraries.
 No longer suppress `@doc`, `@summary`, and `@example` decorators. These can be toggled using the APIView UI.
 Support rendering multi-line strings.
 Change default path for generating artifacts.
 
 ## Version 0.1.1 (10-13-2022)
 Support compiler-level noEmit option.
-Support --output-dir option.
+Support `output-dir` emitter option.
 
 ## Version 0.1.0 (10-5-2022)
 Initial release.