+++ tools/apiview/emitters/cadl-apiview/CHANGELOG.md
@@ -1,8 +1,10 @@
 # Release History
 
 ## Version 0.1.2 (Unreleased)
-Support `service-namespace` emitter option to filter the appropriate namespace when it cannot be automatically resolved.
+Support `namespace` emitter option to filter the appropriate namespace when it cannot be automatically resolved.
   This is primarily intended for creating APIViews for libraries.
+Support `version` emitter option to choose which version of a multi-versioned spec to emit. Specs with a single
+  version can omit this. Multi-version specs can omit this if emitting the latest version.
 No longer suppress `@doc`, `@summary`, and `@example` decorators. These can be toggled using the APIView UI.
 Support rendering multi-line strings.
 Change default path for generating artifacts.